refactor: set controllers to infra
