+++ src/modules/customers/domain/entities/user.entity.ts
