+++ src/modules/groups/domain/entities/group-member.entity.ts
