+++ src/modules/groups/domain/entities/group.entity.ts
