+++ src/modules/groups/infra/database/typeorm/mappers/typeorm-group-member.mapper.ts
@@ -1,4 +1,4 @@
-import { GroupMemberEntity } from 'src/apps/group-service/domain/entities/group-member.entity';
+import { GroupMemberEntity } from 'src/modules/groups/domain/entities/group-member.entity';
 import { TypeormGroupMemberModel } from '../models/typeorm-group-member.model';
 
 export class TypeormGroupMemberMapper {
