+++ src/modules/groups/infra/database/typeorm/mappers/typeorm-group.mapper.ts
@@ -1,4 +1,4 @@
-import { GroupEntity } from 'src/apps/group-service/domain/entities/group.entity';
+import { GroupEntity } from 'src/modules/groups/domain/entities/group.entity';
 import { TypeormGroupModel } from '../models/typeorm-group.model';
 
 export class TypeormGroupMapper {
