+++ src/modules/groups/infra/database/typeorm/models/typeorm-group-member.model.ts
@@ -1,4 +1,4 @@
-import { GroupEntity } from 'src/apps/group-service/domain/entities/group.entity';
+import { GroupEntity } from 'src/modules/groups/domain/entities/group.entity';
 import {
   Column,
   CreateDateColumn,
