+++ src/modules/groups/infra/http/controllers/create-group-controller.ts
@@ -5,7 +5,7 @@ import {
   ApiResponse,
   ApiBadRequestResponse,
 } from '@nestjs/swagger';
-import { CreateGroupUseCase } from '../use-cases/create-group/create-group.use-case';
+import { CreateGroupUseCase } from '../../../application/use-cases/create-group/create-group.use-case';
 import { CreateGroupInputDTO } from '../dtos/create-group.dto';
 
 @Controller('groups')
