+++ src/modules/groups/infra/http/dtos/create-group.dto.ts
